requirements.txt
python-telegram-bot==10.1.0
quart==0.6.4
python-dotenv==0.9.0
pony==0.7.5